backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.21.6
tensorflow-cpu==2.9.0
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.4.2
python-multipart==0.0.6 